Predictive-Maintenance/requirements.txt
flask
numpy
tensorflow
kafka-python
requests
prometheus_client